Reason - Reindexing Is Required at the time of Major Version jumps only.
Reformat the code -> Merge all the code related to Index migration to the file FeatureMigrationService.java
