+++ server/modules/upgrade/src/main/java/co/hyperflex/upgrade/services/migration/FeatureMigrationService.java
@@ -2,10 +2,18 @@ package co.hyperflex.upgrade.services.migration;
 
 import co.hyperflex.clients.client.ApiRequest;
 import co.hyperflex.clients.elastic.ElasticsearchClientProvider;
+import co.hyperflex.core.exceptions.AppException;
+import co.hyperflex.core.exceptions.HttpStatus;
 import co.hyperflex.core.services.upgrade.ClusterUpgradeJobService;
+import co.hyperflex.core.upgrade.ClusterUpgradeJobEntity;
 import co.hyperflex.core.utils.VersionUtils;
+import co.hyperflex.precheck.utils.IndexUtils;
+import co.hyperflex.upgrade.services.dtos.IndexReindexInfo;
+import co.hyperflex.upgrade.services.dtos.MigrationInfoResponse;
+import co.hyperflex.upgrade.services.dtos.ReindexStatus;
 import com.fasterxml.jackson.databind.JsonNode;
 import jakarta.validation.constraints.NotNull;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -15,11 +23,16 @@ public class FeatureMigrationService {
   private static final Logger log = LoggerFactory.getLogger(FeatureMigrationService.class);
   private final ElasticsearchClientProvider elasticsearchClientProvider;
   private final ClusterUpgradeJobService clusterUpgradeJobService;
+  private final IndexUtils indexUtils;
+  private final IndexMigrationService indexMigrationService;
 
   public FeatureMigrationService(ElasticsearchClientProvider elasticsearchClientProvider,
-                                 ClusterUpgradeJobService clusterUpgradeJobService) {
+                                 ClusterUpgradeJobService clusterUpgradeJobService, IndexUtils indexUtils,
+                                 IndexMigrationService indexMigrationService) {
     this.elasticsearchClientProvider = elasticsearchClientProvider;
     this.clusterUpgradeJobService = clusterUpgradeJobService;
+    this.indexUtils = indexUtils;
+    this.indexMigrationService = indexMigrationService;
   }
 
   public FeatureMigrationResponse migrate(String clusterId) {
@@ -33,6 +46,19 @@ public class FeatureMigrationService {
 
   public @NotNull GetFeatureMigrationResponse getFeatureMigrationResponse(String clusterId) {
     var upgradeJob = clusterUpgradeJobService.getLatestJobByClusterId(clusterId);
+    ClusterUpgradeJobEntity job = clusterUpgradeJobService.getActiveJobByClusterId(clusterId);
+
+    String currentVersion = job.getCurrentVersion();
+    String targetVersion = job.getTargetVersion();
+
+    Integer currentMajor = indexUtils.getMajorVersion(currentVersion);
+    Integer targetMajor = indexUtils.getMajorVersion(targetVersion);
+
+    // Reindexing Is Required at the time of Major Version jumps only
+    if (currentMajor + 1 != targetMajor) {
+      return new GetFeatureMigrationResponse(FeatureMigrationStatus.NO_MIGRATION_NEEDED);
+    }
+
     if (VersionUtils.isVersionGte(upgradeJob.getCurrentVersion(), "7.16.0")) {
       var client = elasticsearchClientProvider.getClient(clusterId);
       var response = client.execute(ApiRequest.builder(JsonNode.class).get().uri("/_migration/system_features").build());
@@ -44,4 +70,46 @@ public class FeatureMigrationService {
     }
   }
 
+  public MigrationInfoResponse getMigrationInfo(String clusterId) {
+
+    try {
+      ClusterUpgradeJobEntity upgradeJob = clusterUpgradeJobService.getLatestJobByClusterId(clusterId);
+      String currentVer = upgradeJob.getCurrentVersion();
+      String targetVer = upgradeJob.getTargetVersion();
+
+      boolean isValidUpgradePath = VersionUtils.isValidUpgrade(currentVer, targetVer);
+      List<IndexReindexInfo> reindexNeedingIndices = indexMigrationService.getReindexIndicesMetadata(clusterId);
+      GetFeatureMigrationResponse featureMigrationResponse = getFeatureMigrationResponse(clusterId);
+
+      ReindexStatus status = determineReindexStatus(reindexNeedingIndices, isValidUpgradePath, currentVer, featureMigrationResponse);
+
+      return new MigrationInfoResponse(
+          isValidUpgradePath,
+          featureMigrationResponse,
+          reindexNeedingIndices,
+          status
+      );
+    } catch (Exception e) {
+      log.error("Failed to get migration info for cluster {}", clusterId, e);
+      throw new AppException(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+  }
+
+  private ReindexStatus determineReindexStatus(List<IndexReindexInfo> indices, boolean isValidUpgradePath, String currentVer,
+                                               GetFeatureMigrationResponse featureMigrationResponse) {
+    if (indices.isEmpty()) {
+      return new ReindexStatus(false, "No indices require reindexing");
+    }
+    if (!isValidUpgradePath) {
+      return new ReindexStatus(false, "Currently in view only mode. Select valid target version.");
+    }
+    if (!VersionUtils.isVersionGte(currentVer, "8.18.0")) {
+      return new ReindexStatus(false, "Direct reindexing available from 8.18.0. Manual reindex required.");
+    }
+    if (featureMigrationResponse.status() != FeatureMigrationStatus.NO_MIGRATION_NEEDED) {
+      return new ReindexStatus(false, "First Migrate Features");
+    }
+
+    return new ReindexStatus(true, null);
+  }
 }
