--- server/modules/upgrade/src/main/java/co/hyperflex/upgrade/services/migration/IndexMigrationResponse.java
@@ -1,5 +1,0 @@
-package co.hyperflex.upgrade.services.migration;
-
-public record IndexMigrationResponse() {
-
-}
